Feat : 카테고리 관리 코드 정리
