+++ Cookiee/Cookiee/Presentation/MyPage/Category/View/CategoryListView.swift
@@ -10,10 +10,9 @@ import SwiftUI
 struct CategoryListView: View {
     @Environment(\.presentationMode) var presentationMode: Binding<PresentationMode>
     
-    @StateObject var stateCategoryListViewModel = CategoryListViewModel()
+    @ObservedObject var stateCategoryListViewModel = CategoryListViewModel()
 
     @State private var isAddButtonTapped = false
-    
     @State private var isDeleteButtonTapped = false
     @State private var categoryNameToDelete: String?
     @State private var categoryIdToDelete: String?
@@ -77,6 +76,9 @@ struct CategoryListView: View {
                 stateCategoryListViewModel.loadCategoryListData()
            }
         }
+//        .onChange(of: stateCategoryListViewModel.isUpdateSuccess) {
+//            presentationMode.wrappedValue.dismiss()
+//        }
         .showCustomAlert(
             isPresented: $isDeleteButtonTapped,
             content: {
@@ -107,108 +109,11 @@ struct CategoryListView: View {
                     title: Text("삭제하기").foregroundColor(Color.Brown00))
         )
     }
+    
+    
 }
 
 #Preview {
     CategoryListView()
 }
 
-// MARK: - 카테고리 리스트 (수정 버튼, 삭제 버튼)
-struct CategoryListRowView: View {
-    @State var id: String = ""
-    @State var name: String = ""
-    @State var color: String
-    @State private var isEditButtonTapped = false
-    
-    @Binding var isDeleteButtonTapped: Bool
-    @Binding var categoryNameToDelete: String?
-    @Binding var categoryIdToDelete: String?
-    
-    @ObservedObject var categoryListViewModel : CategoryListViewModel
-    
-    var body: some View {
-        VStack {
-            HStack {
-                // 색상
-                Rectangle()
-                    .fill(Color(hex: color))
-                    .frame(width: 25, height: 25)
-                    .cornerRadius(3.0)
-                
-                // 이름
-                HStack {
-                    Text(name)
-                    Spacer()
-                    // 수정 버튼
-                    Button(action: {
-                        isEditButtonTapped = true
-                    }, label: {
-                        Image("EditIconBrown")
-                    })
-                }
-                .padding(10)
-                .font(.Body1_M)
-                .frame(height: 35)
-                .background(Color.Gray01)
-                .cornerRadius(3.0)
-                            
-                // 삭제 버튼
-                Button(action: {
-                    categoryNameToDelete = name
-                    categoryIdToDelete = id
-                    isDeleteButtonTapped = true
-                }, label: {
-                    Image("TrashIconRed")
-                })
-            }
-            Divider()
-        }
-        .sheet(isPresented: $isEditButtonTapped, onDismiss: {
-            print("카테고리 수정 onDismiss")
-            categoryListViewModel.loadCategoryListData()
-            
-        }) {
-            CategoryAddAndEditView(
-                categoryListViewModel: categoryListViewModel,
-                isNewCategory: false,
-                id: id,
-                name: name,
-                selectedColor: color,
-                toggleIsOpenCategoryAddSheet: {
-                isEditButtonTapped.toggle()
-            })
-                .presentationDetents([.fraction(0.95)])
-                .presentationDragIndicator(Visibility.visible)
-        }
-    }
-}
-
-// MARK: - 카테고리 추가 버튼
-struct CategoryAddButtonView: View {
-    @State var name: String = ""
-    @State var toggleIsTapped: () -> Void
-    
-    var body: some View {
-        VStack {
-            HStack {
-                Rectangle()
-                    .fill(Color.Gray02)
-                    .frame(width: 25, height: 25)
-                    .cornerRadius(/*@START_MENU_TOKEN@*/3.0/*@END_MENU_TOKEN@*/)
-                    .overlay(Image("Plus"))
-                HStack {
-                    Button(action: {
-                        toggleIsTapped()
-                    }, label: {
-                        Text("추가하기")
-                            .font(.Body1_M)
-                            .foregroundColor(.black)
-                    })
-                }
-                .font(.Body1_M)
-                .frame(height: 35)
-                Spacer()
-            }
-        }
-    }
-}
