+++ Cookiee/Cookiee/Presentation/MyPage/Category/ViewModel/CategoryListViewModel.swift
@@ -7,33 +7,44 @@
 
 import Foundation
 
+//struct CategoryForListRowDTO : Identifiable {
+//    let categoryId: Int64
+//    let categoryName: String
+//    let categoryColor: String
+//    
+//    var id: Int64 {
+//        return categoryId
+//    }
+//}
+
 class CategoryListViewModel: ObservableObject {
     @Published var categories: [CategoryResultData] = []
-//    {
-//        willSet(new) {
-//            print("🔥 이 카테고리 값으로 변경될 예정 \(new)")
-//        }
-//        didSet {
-//            print("🔥 didSet 작동함")
-//            loadCategoryListData()
-//          }
-//    }
+    @Published var isUpdateSuccess: Bool = false
+
     let categoryService = CategoryService()
 
     func loadCategoryListData() {
-        categoryService.getCategoryList() { result in
+        categoryService.getCategoryList { result in
             switch result {
             case .success(let categoryList):
                 DispatchQueue.main.async {
-                    self.categories = categoryList.result
+                    self.categories = categoryList.result.map {
+                        CategoryResultData(
+                            categoryId: $0.categoryId,
+                            categoryName: $0.categoryName,
+                            categoryColor: $0.categoryColor
+                        )
+                    }
                     print("✅ loadCategoryListData 성공")
                     print("🍎 loadCategoryListData 결과 : ", self.categories)
                 }
             case .failure(let error):
-                print(error)
+                print("❌ Error loading categories: \(error)")
             }
         }
     }
+
+
     
     func addCategory(categoryName: String, categoryColor: String) {
         let categoryRequest = CategoryRequestDTO(categoryName: categoryName, categoryColor: categoryColor)
@@ -42,7 +53,6 @@ class CategoryListViewModel: ObservableObject {
             switch result {
             case .success(let categoryResponse):
                 DispatchQueue.main.async {
-                    self.objectWillChange.send()
                     self.loadCategoryListData()
                     print("✅ addCategory 성공")
                     print("🍎 addCategory 결과 : ", categoryResponse)
@@ -55,30 +65,25 @@ class CategoryListViewModel: ObservableObject {
     }
     
     func updateCategory(categoryId: String, categoryName: String, categoryColor: String) {
-        let categoryRequest = CategoryRequestDTO(categoryName: categoryName, categoryColor: categoryColor)
-        
-        categoryService.putCategory(cateId: categoryId.description, requestBody: categoryRequest) { result in
-            switch result {
-            case .success(let categoryResponse):
-                DispatchQueue.main.async {
-                    self.objectWillChange.send()
-                    self.loadCategoryListData()
-                    print("✅ updateCategory 성공")
-                    print("🍎 updateCategory 결과 : ", categoryResponse)
-               }
-
-            case .failure(let error):
-                print("updateCategory error:", error)
+            let categoryRequest = CategoryRequestDTO(categoryName: categoryName, categoryColor: categoryColor)
+            categoryService.putCategory(cateId: categoryId.description, requestBody: categoryRequest) { result in
+                switch result {
+                case .success:
+                    DispatchQueue.main.async {
+                        self.isUpdateSuccess = true
+                        self.loadCategoryListData()
+                    }
+                case .failure(let error):
+                    print("updateCategory error:", error)
+                }
             }
         }
-    }
     
     func removeCategory(categoryId: String) {
         categoryService.deleteCategory(cateId: categoryId.description) { result in
             switch result {
             case .success(let categoryResponse):
                 DispatchQueue.main.async {
-                    self.objectWillChange.send()
                     self.loadCategoryListData()
                     print("✅ removeCategory 성공")
                     print("🍎 removeCategory 결과 : ", categoryResponse)
